Create dye_script
This was used for checking 1uL pipetting using bromophenol blue dye and water. All work performed using an opentrons p20 pipette, opentrons p20 pipette tips, and a 96-well plate.

User must input coordinate locations of the first tip in the tip box, dye well, and first sample well, as well as the number of replicates for each draw and number of samples. 

The Kraken will pick up a tip, go to the specified well and collect 20uL dye. It will then go to the first sample well and dispense 1uL in subsequent wells as defined by "replicate_number". It will discard the tip and go to collect a new tip, repeating as defined by "sample_number".