Update config.py
***Hard Coded Movement Parameters***

ALL LOCATIONS IN OUR CODE ARE DEFINED RELATIVE TO THE TOP RIGHT CORNER OF THE SLOT NEAREST (0,0). 

As defined for our particular machine, these dictate all movement parameters for the Kraken. For any new machine all locations should be confirmed. For any new object the distance between functionality A1 and the top right corner of the slot (closest to origin) should be determined, along with functionality separation and arrangement. CORRECT CLEARANCE HEIGHT DETERMINATION IS ESSENTIAL - INCORRECT CLEARANCE HEIGHT CAN RESULT IN COLLISIONS BETWEEN THE TOOL AND AN OBJECT.